requirements.txt
boto3==1.42.53
botocore==1.42.53
jmespath==1.1.0
load-dotenv==0.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.21.0
pydantic==2.12.5
pydantic_core==2.41.5
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
